+++ app/models/order.rb
@@ -6,6 +6,7 @@ class Order < ApplicationRecord
   belongs_to :approver_admin_user, class_name: 'AdminUser', optional: true
   has_and_belongs_to_many :products
   has_many :payments
+  has_many :subscriptions
 
   after_create :set_amounts!
   #after_update :update_paid_amounts!
@@ -27,7 +28,7 @@ class Order < ApplicationRecord
     end
 
     event :complete do
-      transitions from: [:just_made, :processing], to: :completed
+      transitions from: [:just_made, :processing], to: :completed, success: :create_subscriptions
     end
 
     event :cancel do
@@ -55,9 +56,17 @@ class Order < ApplicationRecord
   def set_state!
     if self.amount_paid_cents == self.amount_to_pay_cents
       #TODO create subscriptions from purchaded products
-      self.complete!
+      self.complete! if may_complete?
     else
       self.start_process! if self.may_start_process?
+    end
+  end
+
+  # Move it in a service if this method become more complex
+  def create_subscriptions
+    products.each do |product|
+      subscriptions << Subscription.new(organization: organization, category_id: product.category_id, product: product,
+                                        user: user, start_on: Time.zone.today, end_on: Time.zone.today + product.days)
     end
   end
 end