+++ app/models/payment.rb
@@ -5,12 +5,24 @@ class Payment < ApplicationRecord
   belongs_to :user, optional: true
   belongs_to :order, optional: true
 
-  after_create :update_order!, if: :order
+  after_create :add_to_order!, if: :order
+  after_update :update_order!, if: :order
 
   enum source: { stripe: 0, paypal: 1, cc: 2 }, _prefix: true
 
   STATES = { just_made: 10, canceled: 30, confirmed: 40}
   enum state: STATES
+
+  def confirm!
+    update! state: :confirmed
+  end
+
+  private
+
+  def add_to_order!
+    return if order.blank? || !confirmed?
+    order.update_paid_amounts!
+  end
 
   def update_order!
     return unless order