+++ db/migrate/20200304182620_create_orders.rb
@@ -9,6 +9,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
       t.integer :amount_paid_cents, limit: 4, default: 0
       t.integer :discount_cents, limit: 4, default: 0
       t.integer :currency, limit: 1, default: 0
+      t.date :start_on
       t.datetime :paid_at
       t.references :approver_admin_user, null: true, foreign_key: {to_table: :admin_users}