+++ spec/factories/orders.rb
@@ -3,5 +3,6 @@ FactoryBot.define do
     organization
     user { build(:user, organization: organization) }
     sequence(:total_amount_cents, 10000) { |n| (n+1) * 100 }
+    start_on { Time.zone.today }
   end
 end